requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.2
yfinance==0.2.32
requests==2.31.0
pyyaml==6.0.1
